\name{ATESurvival-package}
\alias{ATESurvival-package}
\alias{ATESurvival}
\docType{package}
\title{
  Computation of the ATE and bootstrapped confidence intervals/bands for time-to-event data with competing risks
}
\description{
  Confidence intervals/bands can be determined using influence functions, the wild bootstrap, and the classical bootstrap.
}
\details{
  The function 'ATE_IF_WBS' computes the ATE and the associated confidence intervals/bands using influence functions or the wild bootstrap.
  The function 'EBS' computes the ATE for data where the observations were drawn with replacement.
}
\author{
Jasmin Ruehl <jasmin.ruehl@math.uni-augsburg.de>
}
\keyword{ package }
